chromium: Use host-gn and use a custom GN toolchain file
Instead of using the ChromeOS toolchain definitions, which is an ugly
hack, the build recipe adds a GN snippet to the souce tree which points
directly to the Buildroot toolchain.

Also, tweaked the build options a bit, and added handling of a few
Buildroot global options (e.g. debugging levels) so the chosen options
get passed down to the Chromium build.