chromium: Simplify configuration, build, and installation commands
The expansions of $(CHROMIUM_NINJA) and $(CHROMIUM_GN) are used only
once, so move them to their corresponding command lists.

The commands for installing can be greatly simplified using "install".
Also, the "&& \" trailers for each commands only make the single command
passed to the shell longer, when it is perfectly fine to have multiple
commands in CHROMIUM_INSTALL_TARGET_CMDS.