wpebackend-rdk: Fix touch input crashes
Added two downstream patches. The first one fixes the crash that
happens when touch input is provided, but it ignores the input.
The second patch fakes touch events as mouse events so they can be
processed by WebKit.

This is a temporary solution while the problem is not properly fixed
upstream. It must be explicitly enabled in runtime with the envvar
WPE_TOUCH_FAKE_MOUSE=1.